+++ lib/Lime/Helper/Session.php
@@ -15,14 +15,14 @@ class Session extends \Lime\Helper {
 
         if ($this->initialized) return;
 
-        if (\session_status() != PHP_SESSION_ACTIVE) {
+        if (session_status() != PHP_SESSION_ACTIVE) {
 
             $this->name = $name ? $name : $this->app->retrieve('session.name');
 
-            \session_name($this->name);
-            \session_start();
+            session_name($this->name);
+            session_start();
         } else {
-            $this->name = \session_name();
+            $this->name = session_name();
         }
 
         $this->initialized = true;
@@ -42,14 +42,26 @@ class Session extends \Lime\Helper {
     }
 
     public function destroy(): void {
-        \session_destroy();
+        session_destroy();
     }
 
     public function close(): void {
-        \session_write_close();
+        session_write_close();
     }
 
     public function regenerateId(bool $delete_old_session = false): bool {
-        return \session_regenerate_id($delete_old_session);
+
+        if (session_status() != PHP_SESSION_ACTIVE) {
+            return false;
+        }
+
+        if ($delete_old_session) {
+
+            foreach ($this->session as $key => $value) {
+                unset($this->session[$key]);
+            }
+        }
+
+        return session_regenerate_id($delete_old_session);
     }
 }