+++ modules/App/Controller/Auth.php
@@ -20,15 +20,16 @@ class Auth extends Base {
 
         $redirectTo = $this->param('to', '/');
 
-        if (\substr($redirectTo, 0, 1) !== '/') {
+        if (substr($redirectTo, 0, 1) !== '/') {
             $redirectTo = '/';
         }
 
         $redirectTo = htmlspecialchars($this->app->routeUrl($redirectTo), ENT_QUOTES, 'UTF-8');
+        $csrfToken = $this->helper('csrf')->token('app.login');
 
         $this->helper('theme')->pageClass('login-page');
 
-        return $this->render('app:views/auth/login.php', \compact('redirectTo'));
+        return $this->render('app:views/auth/login.php', compact('redirectTo', 'csrfToken'));
     }
 
     public function logout() {
@@ -50,11 +51,11 @@ class Auth extends Base {
 
         $auth = $this->param('auth');
 
-        if (!$auth || !isset($auth['user'], $auth['password']) || !\is_string($auth['user']) || !\is_string($auth['password'])) {
+        if (!$auth || !isset($auth['user'], $auth['password']) || !is_string($auth['user']) || !\is_string($auth['password'])) {
             return $this->stop(412);
         }
 
-        if (!$this->helper('csrf')->isValid('login', $this->param('csrf'), true)) {
+        if (!$this->helper('csrf')->isValid('app.login', $this->param('csrf'), true)) {
             return $this->stop(412);
         }
 
@@ -70,7 +71,7 @@ class Auth extends Base {
 
         $user = $this->helper('auth')->authenticate($auth);
 
-        if ($user && $user['role'] == 'public') {
+        if ($user && $user['role'] === 'public') {
             return $this->stop(412);
         }