+++ modules/App/Helper/Csrf.php
@@ -4,9 +4,23 @@ namespace App\Helper;
 
 class Csrf extends \Lime\Helper {
 
+    protected ?string $sessionKey;
+
+    protected function initialize() {
+
+        $key = $this->app->helper('session')->read("app.csrf._key", null);
+
+        if (!$key) {
+            $key = bin2hex(random_bytes(32));
+            $this->app->helper('session')->write("app.csrf._key", $key);
+        }
+
+        $this->sessionKey = $key;
+    }
+
     public function generateToken(string $key, ?int $expire = null): string {
 
-        $payload = ['csrf' => $key];
+        $payload = ['csrf' => "{$key}.{$this->sessionKey}"];
 
         if ($expire && is_numeric($expire)) {
             $payload['exp'] = $expire;
@@ -37,9 +51,13 @@ class Csrf extends \Lime\Helper {
         }
 
         if ($checkpayload) {
+
             try {
+
                 $payload = $this->app->helper('jwt')->decode($token);
-                return isset($payload->csrf) && $payload->csrf == $key;
+
+                return isset($payload->csrf) && $payload->csrf === "{$key}.{$this->sessionKey}";
+
             } catch(\Exception $e) {
                 return false;
             }