+++ modules/App/layouts/app.php
@@ -6,8 +6,15 @@ $sidePanelContents = $this->block('app-side-panel', ['print' => false]);
 
 ?>
 <!DOCTYPE html>
-<html lang="en" class="<?= $this->helper('theme')->pageClass() ?>" data-base="<?= rtrim($this->baseUrl('/'), '/') ?>" data-route="<?= rtrim($this->routeUrl('/'), '/') ?>" data-csrf="<?= $this->helper('csrf')->token('app-csrf') ?>" data-version="<?= $this->retrieve('app.version') ?>" data-theme="<?= $this->helper('theme')->theme() ?>">
-
+<html
+    lang="en"
+    class="<?= $this->helper('theme')->pageClass() ?>"
+    data-base="<?= rtrim($this->baseUrl('/'), '/') ?>"
+    data-route="<?= rtrim($this->routeUrl('/'), '/') ?>"
+    data-csrf="<?= $this->helper('csrf')->token('app.csrf') ?>"
+    data-version="<?= $this->retrieve('app.version') ?>"
+    data-theme="<?= $this->helper('theme')->theme() ?>"
+>
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">